requirements.txt
black
bump2version
click
contextvars; python_version < "3.7"
coverage[toml]
dataclasses; python_version < "3.7"
dephell[full]
filelock
fissix
flake8
immutables
ipython
isort[pyproject]
jsondiff
loguru
mmdnn
more-itertools
mypy
mypy-protobuf
numpy
onnx
onnxruntime
pip>=19.3
pre-commit
protobuf
pytest
pytest-xdist
pyyaml
scipy
sphinx
tensorflow==1.15.3
torch==1.6.0
torchvision==0.7.0
tox
twine
typing; python_version < "3.7"
typing-extensions
watchdog
wget
wheel
